Fix events

+++ src/ScriptDotNet.Abstract/Events/CharAnimationEventArgs.cs
@@ -2,12 +2,12 @@
 {
     public class CharAnimationEventArgs : ObjectEventArgs
     {
-        public CharAnimationEventArgs(uint objectId, uint action)
+        public CharAnimationEventArgs(int objectId, ushort action)
             : base(objectId)
         {
             Action = action;
         }
-        public uint Action { get; private set; }
+        public ushort Action { get; private set; }
     }
