+++ src/ScriptDotNet.Abstract/Events/SoundEventArgs.cs
@@ -4,7 +4,7 @@ namespace ScriptDotNet
 {
     public class SoundEventArgs : EventArgs
     {
-        public SoundEventArgs(ushort soundId, ushort x, ushort y, int z)
+        public SoundEventArgs(ushort soundId, ushort x, ushort y, short z)
         {
             SoundId = soundId;
             X = x;
@@ -14,7 +14,7 @@ namespace ScriptDotNet
         public ushort SoundId { get; private set; }
         public ushort X { get; private set; }
         public ushort Y { get; private set; }
-        public int Z { get; private set; }
+        public short Z { get; private set; }
     }
